+++ packages/parser/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.44](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.43...@vtj/parser@0.12.44) (2025-06-20)
+
+
+### Bug Fixes
+
+* 🐛 openApi ([d29d618](https://gitee.com/newgateway/vtj/commits/d29d618dc071f21ac3042b6b0ba305ac68fffc55))
+
+
+
+
+
+## [0.12.43](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.42...@vtj/parser@0.12.43) (2025-06-17)
+
+**Note:** Version bump only for package @vtj/parser
+
+
+
+
+
+## [0.12.42](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.41...@vtj/parser@0.12.42) (2025-06-17)
+
+
+### Bug Fixes
+
+* 🐛 parser template v-if v-for 文本节点或多个元素节点 ([b154964](https://gitee.com/newgateway/vtj/commits/b154964eb6e8403b2a0c2c0d9bda3605ebccd435))
+
+
+
+
+
 ## [0.12.41](https://gitee.com/newgateway/vtj/compare/@vtj/parser@0.12.40...@vtj/parser@0.12.41) (2025-06-16)
 
 **Note:** Version bump only for package @vtj/parser